Move tx status
